put stuff that parseNewActivity and parseUpdatedActivity have in common into functions
